Use hashed filenames when saving
Refactor File.getName() method to return a string consisting of the
hashed filename, a random string and the current timestamp.

This way, there should be no more issue with unsupported characters in the
filename.

Furthermore, Helma’s internal access names are now fully decoupled from
the filename and thus, the full unicode set should be available for naming.